# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShiftGuard AI API
  version: 1.1.0
  summary: Integration contract for the ShiftGuard AI scheduling MVP
  description: |
    The ShiftGuard AI API manages employees and availability, generates advisory
    staffing recommendations, and records a manager's final decision.

    AI output is advisory only. Every recommendation is created as a draft and
    must be explicitly approved or rejected by a manager.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:5000
    description: Local Flask development server
...
paths:
...
  /api/employees:
    get:
      tags: [Employees]
      operationId: listEmployees
      summary: List employees
      description: Returns all employees, including inactive employees, ordered by name.
      responses:
        "200":
          description: Employee collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      tags: [Employees]
      operationId: createEmployee
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
            example:
              name: Alex Rivera
              role: Nurse
              max_weekly_hours: 40
              hourly_rate: 30
      responses:
        "201":
          description: Employee created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: An employee with the same name already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: An employee with that name already exists
  /api/employees/{employee_id}:
    parameters:
      - $ref: "#/components/parameters/EmployeeId"
    patch:
      tags: [Employees]
      operationId: updateEmployee
      summary: Update or deactivate an employee
      description: |
        Updates one or more employee fields. Set `active` to false to deactivate
        an employee without deleting historical schedules or availability.
        Deactivated employees are excluded from new recommendations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
            example:
              role: Supervisor
              max_weekly_hours: 35
              active: false
      responses:
        "200":
          description: Updated employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: The employee does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Employee not found
        "409":
          description: Another employee already uses the requested name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/employees/{employee_id}/availability:
    parameters:
      - $ref: "#/components/parameters/EmployeeId"
    get:
      tags: [Employees]
      operationId: listEmployeeAvailability
      summary: List an employee's recurring availability
      responses:
        "200":
          description: Availability ordered by day and start time.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityListResponse"
        "404":
          description: The employee does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Employee not found
    post:
      tags: [Employees]
      operationId: addEmployeeAvailability
      summary: Add recurring weekly availability
      description: |
        `day_of_week` is zero-based: 0 is Monday and 6 is Sunday. The end
        time must be later than the start time; overnight ranges are not supported.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilityRequest"
            example:
              day_of_week: 1
              start_time: "08:00"
              end_time: "18:00"
      responses:
        "201":
          description: Availability created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: The employee does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Employee not found
        "409":
          description: The same availability record already exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: That availability record already exists
  /api/availability/{availability_id}:
    parameters:
      - $ref: "#/components/parameters/AvailabilityId"
    patch:
      tags: [Employees]
      operationId: updateAvailability
      summary: Update recurring availability
      description: Updates one or more fields while preserving omitted values.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilityRequest"
            example:
              start_time: "07:30"
              end_time: "17:30"
      responses:
        "200":
          description: Updated availability.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/AvailabilityNotFound"
        "409":
          description: The update would duplicate another availability record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: That availability record already exists
    delete:
      tags: [Employees]
      operationId: deleteAvailability
      summary: Delete recurring availability
      responses:
        "204":
          description: Availability deleted; the response has no body.
        "404":
          $ref: "#/components/responses/AvailabilityNotFound"
  /api/shifts:
    get:
      tags: [Scheduling]
      operationId: listShifts
      summary: List and filter schedules
      description: |
        Returns schedules in ascending date, start-time, and identifier order.
        Filters are combined with AND. `employee_id` matches schedules containing
        an assignment for that employee.
      parameters:
        - name: date_from
          in: query
          description: Inclusive start date.
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: Inclusive end date.
          schema:
            type: string
            format: date
        - name: required_role
          in: query
          description: Case-insensitive exact role match.
          schema:
            type: string
            minLength: 1
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ShiftStatus"
        - name: employee_id
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Matching schedule collection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/shifts/recommendations:
    post:
      tags: [Scheduling]
      operationId: createShiftRecommendation
      summary: Generate and store a draft schedule recommendation
      description: |
        Predicts a staffing level unless `required_staff` is supplied, filters
        employees by role and availability, excludes projected overtime by
        default, ranks eligible employees, and stores the result as a draft.

        Supplying `required_staff` is a manager override and results in
        `model_source: manager_override`. Setting `allow_overtime` to true makes
        overtime candidates eligible; it does not approve the draft.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecommendationRequest"
            examples:
              predictedStaffing:
                summary: Let the model choose the staffing level
                value:
                  shift_date: "2026-09-01"
                  start_time: "09:00"
                  end_time: "17:00"
                  required_role: Nurse
                  workload_score: 80
              managerOverride:
                summary: Manager supplies the staffing level
                value:
                  shift_date: "2026-09-01"
                  start_time: "09:00"
                  end_time: "17:00"
                  required_role: Nurse
                  workload_score: 60
                  required_staff: 2
                  allow_overtime: false
      responses:
        "201":
          description: Draft recommendation created; manager action is still required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/shifts/{shift_id}:
    parameters:
      - $ref: "#/components/parameters/ShiftId"
    get:
      tags: [Scheduling]
      operationId: getShift
      summary: Retrieve a schedule recommendation
      responses:
        "200":
          description: Draft, approved, or rejected schedule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "404":
          $ref: "#/components/responses/ShiftNotFound"
  /api/shifts/{shift_id}/decision:
    parameters:
      - $ref: "#/components/parameters/ShiftId"
    patch:
      tags: [Scheduling]
      operationId: decideShift
      summary: Approve or reject a draft recommendation
      description: |
        A recommendation may be decided only once. The decision is also applied
        to every assignment in the recommendation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecideShiftRequest"
            example:
              decision: approved
              manager_name: Demo Manager
              manager_note: Availability confirmed with the team.
      responses:
        "200":
          description: Decision recorded and updated schedule returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/ShiftNotFound"
        "409":
          description: The schedule has already been approved or rejected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Only draft recommendations can be decided
components:
  parameters:
    EmployeeId:
      name: employee_id
      in: path
      required: true
      description: Positive employee identifier.
      schema:
        type: integer
        minimum: 1
    ShiftId:
      name: shift_id
      in: path
      required: true
      description: Positive shift identifier.
      schema:
        type: integer
        minimum: 1
    AvailabilityId:
      name: availability_id
      in: path
      required: true
      description: Positive availability identifier.
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Invalid JSON, a missing field, or a failed validation rule.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidBody:
              value:
                error: Request body must be a JSON object
            invalidTime:
              value:
                error: Overnight shifts are not supported in this MVP
    ShiftNotFound:
      description: The shift does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Shift not found
    AvailabilityNotFound:
      description: The availability record does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Availability not found
  schemas:
...
    EmployeeListResponse:
      type: object
      required: [employees]
      properties:
        employees:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
    CreateEmployeeRequest:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
          minLength: 1
          description: Unique employee name after surrounding whitespace is removed.
        role:
          type: string
          minLength: 1
          description: Role matched case-insensitively during scheduling.
        max_weekly_hours:
          type: number
          minimum: 1
          maximum: 168
          default: 40
        hourly_rate:
          type: number
          minimum: 0
          default: 0
    UpdateEmployeeRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        max_weekly_hours:
          type: number
          minimum: 1
          maximum: 168
        hourly_rate:
          type: number
          minimum: 0
        active:
          type: boolean
          description: False deactivates and true reactivates the employee.
    Employee:
      type: object
      required:
        - id
        - name
        - role
        - max_weekly_hours
        - hourly_rate
        - active
        - created_at
      properties:
        id:
          type: integer
          minimum: 1
        name:
          type: string
        role:
          type: string
        max_weekly_hours:
          type: number
        hourly_rate:
          type: number
        active:
          type: integer
          enum: [0, 1]
          description: SQLite-backed active flag; 1 is active and 0 is inactive.
        created_at:
          type: string
          description: SQLite UTC timestamp in `YYYY-MM-DD HH:MM:SS` form.
          pattern: '^\\d{4}-\\d{2}-\\d{2} \\d{2}:\\d{2}:\\d{2}$'
    CreateAvailabilityRequest:
      type: object
      required: [day_of_week, start_time, end_time]
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
          description: 0 is Monday and 6 is Sunday.
        start_time:
          $ref: "#/components/schemas/LocalTime"
        end_time:
          $ref: "#/components/schemas/LocalTime"
    UpdateAvailabilityRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
        start_time:
          $ref: "#/components/schemas/LocalTime"
        end_time:
          $ref: "#/components/schemas/LocalTime"
    Availability:
      type: object
      required: [id, employee_id, day_of_week, start_time, end_time]
      properties:
        id:
          type: integer
          minimum: 1
        employee_id:
          type: integer
          minimum: 1
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
        start_time:
          $ref: "#/components/schemas/LocalTime"
        end_time:
          $ref: "#/components/schemas/LocalTime"
    AvailabilityListResponse:
      type: object
      required: [availability]
      properties:
        availability:
          type: array
          items:
            $ref: "#/components/schemas/Availability"
    CreateRecommendationRequest:
      type: object
      required:
        - shift_date
        - start_time
        - end_time
        - required_role
        - workload_score
      properties:
        shift_date:
          type: string
          format: date
        start_time:
          $ref: "#/components/schemas/LocalTime"
        end_time:
          $ref: "#/components/schemas/LocalTime"
        required_role:
          type: string
          minLength: 1
        workload_score:
          type: number
          minimum: 0
          maximum: 100
        required_staff:
          type: integer
          minimum: 1
          maximum: 50
          description: Optional manager override for the predicted staffing level.
        allow_overtime:
          type: boolean
          default: false
    DecideShiftRequest:
      type: object
      required: [decision, manager_name]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
          description: Input is normalized to lowercase by the API.
        manager_name:
          type: string
          minLength: 1
        manager_note:
          type: [string, "null"]
...
    ShiftListResponse:
      type: object
      required: [shifts]
      properties:
        shifts:
          type: array
          items:
            $ref: "#/components/schemas/Shift"
    RecommendationResponse:
      allOf:
        - $ref: "#/components/schemas/Shift"
        - type: object
          required: [training_records, coverage_gap]
          properties:
            training_records:
              type: integer
              minimum: 0
              description: Historical records available when this recommendation was made.
            coverage_gap:
              type: integer
              minimum: 0
              description: Required positions for which no eligible employee was selected.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
    LocalTime:
      type: string
      pattern: '^([01]\\d|2[0-3]):[0-5]\\d$'
      example: "09:00"
      description: Local wall-clock time in 24-hour `HH:MM` form.
...
    ShiftStatus:
      type: string
      enum: [draft, approved, rejected]